requirements.txt
Django==5.0.4
djangorestframework==3.15.1
psycopg2==2.9.9
drf-spectacular==0.27.2
django-cors-headers==4.3.1
python-dotenv==1.0.1

# move to dev
flake8==7.0.0